SequenceColor extension - ensureCustomProperties
